requirements.txt
boto3>=1.36.2
qdrant-client>=1.7.0
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.2
aiohttp>=3.9.1
